ping: fix ping rtt time may become minus value in multi-thread scenario
clib_cpu_time_now use rdtsc to get current timestamp, but rdtsc's value
does not synchronize between the CPUs, leading to ping RTT being negative.

Solution: use a thread-local vlib_time_now(vm) for the calling thread.

Type: fix

+++ src/plugins/ping/ping.h
@@ -74,7 +74,7 @@ extern ping_main_t ping_main;
 typedef CLIB_PACKED (struct {
   u16 id;
   u16 seq;
-  u64 time_sent;
+  f64 time_sent;
   u8 data[0];
 }) icmp46_echo_request_t;
